requirements.txt
requests>=2.28.0
Pillow>=9.0.0
torch>=1.9.0
numpy>=1.19.0